+++ src/Application/Query/RegistrationPath/QuestionView.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace Proximum\Vimeet\Application\Query\RegistrationPath;
+
+class QuestionView
+{
+    /** @var int */
+    public $id;
+
+    /** @var string */
+    public $title;
+
+    /** @var array */
+    public $answerViews;
+
+    /**
+     * @param int          $id
+     * @param string       $title
+     * @param AnswerView[] $answerViews
+     */
+    public function __construct(int $id, string $title, array $answerViews)
+    {
+        $this->id = $id;
+        $this->title = $title;
+        $this->answerViews = $answerViews;
+    }
+}