+++ src/Domain/Repository/RegistrationPath/QuestionRepositoryInterface.php
@@ -1,0 +1,31 @@
+<?php
+
+/*
+ * This file is part of the Proximum Vimeet project.
+ *
+ * Copyright (C) Proximum
+ *
+ * @author Elao <contact@elao.com>
+ */
+
+namespace Proximum\Vimeet\Domain\Repository\RegistrationPath;
+
+use Proximum\Vimeet\Domain\Model\Event;
+use Proximum\Vimeet\Domain\Model\RegistrationPath\Question;
+
+interface QuestionRepositoryInterface
+{
+    public function add(Question $question): void;
+
+    public function set(Question $question): void;
+
+    public function remove(Question $question): void;
+
+    /**
+     * @param Event  $event
+     * @param string $locale
+     *
+     * @return Question[]
+     */
+    public function getQuestionsByEvent(Event $event, string $locale): array;
+}