+++ src/Domain/Type/RegistrationPath/View/AnswerView.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Proximum\Vimeet\Domain\Type\RegistrationPath\View;
+
+class AnswerView
+{
+    /**
+     * @var array
+     * example: ['fr' => 'Ma réponse', 'en' => 'My answer']
+     */
+    public $translatedTitle = [];
+}