+++ CHANGELOG.md
@@ -4,9 +4,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
 - VIMEET-2031 - Access control can be enabled or disabled
+- VIMEET-2028 - Configure a registration path
 
-## [2.14.1] - 2019-11-08
+## [2.14.2] - 2019-11-26
+### Fixed
+- MV-208 - Do not close badge when printing
+
+## [2.14.1] - 2019-11-19
 ### Fixed
 - MV-205 - Fix contacts indicators